requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=./requirements.txt ./requirements.in
#
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==7.1.2
    # via uvicorn
fastapi==0.68.0
    # via -r ./requirements.in
future-fstrings==1.2.0
    # via pyecore
h11==0.9.0
    # via uvicorn
idna==3.4
    # via requests
influxdb==5.3.1
    # via -r ./requirements.in
jinja2==3.0.1
    # via -r ./requirements.in
lxml==4.9.1
    # via pyecore
markupsafe==2.1.1
    # via jinja2
msgpack==1.0.4
    # via influxdb
ordered-set==4.1.0
    # via pyecore
overrides==6.1.0
    # via -r ./requirements.in
paho-mqtt==1.6.1
    # via -r ./requirements.in
protobuf==3.19
    # via -r ./requirements.in
pydantic==1.9.0
    # via
    #   -r ./requirements.in
    #   fastapi
pyecore==0.12.1
    # via pyesdl
pyesdl==22.11.2
    # via -r ./requirements.in
python-dateutil==2.8.2
    # via influxdb
python-dotenv==0.20.0
    # via -r ./requirements.in
python-multipart==0.0.5
    # via -r ./requirements.in
pytz==2022.5
    # via influxdb
requests==2.28.1
    # via influxdb
restrictedpython==5.2
    # via pyecore
six==1.16.0
    # via
    #   influxdb
    #   python-dateutil
    #   python-multipart
starlette==0.14.2
    # via fastapi
typing-extensions==4.3.0
    # via pydantic
typing-utils==0.1.0
    # via
    #   -r ./requirements.in
    #   overrides
urllib3==1.26.12
    # via requests
uvicorn==0.11.3
    # via -r ./requirements.in
websockets==8.1
    # via uvicorn
